Fix instance and metadata dependencies
There were times when the controller could be created before the
metadata was, which would cause a setup failure. Similarly, a login
node could be created concurrently with the controller and fail setup
due to NFS mounts.

Soft dependencies were introduced to help terraform create resources in
the correct order while not force replacing on delta (depends_on would
not play nice).