Big TPUs update
In this commit jobsubmit.lua file is introduced to take care of setting
the --switches and fix the -N of the tpu nodes taht contains more than
one vm, the topology.conf is generated for the tpu nodeset, taking into
account the vmcount that they have, this is done to ensure that a job
that is using a TPU group is always contained in it. slurmsync is also
fixed to take into account the tpu nodes. topology/tree is activated for
all the nodes, as the regular nodes also used them for the placemnet
groups. Exclusive is set on the partition level if there are TPU nodes
in this partition. resume.py now chunks the TPU nodes to be resumed
based on the vmcount of the tpunodeset, so that the start_tpu function
can receive a list of slurm_nodes that will be mapped to the worker
nodes of the TPU, this mapping is written in the instance metadata of
the TPU node, that will later be pared by the startup script in order to
set the slurm nodes to the docker container accordingly. jobsubmit lua
calls util.py in order to get the vmcount for a given partition, in
order to do this, now util.py can ba called using the flag -p and
passing it a partition list. The delete Tpu nodes has also benn modified
in order to take into account the "phantom" TPU nodes, these are nodes
that are not mapped to a TPU node directly, but are workers of it. This
is aslo taken into account in the slurmsync so that those nodes are not
unintentionally put down.