fix: not suspending TPU nodes if no regular ones are present
In suspend.py the function to suspend tpu nodes was called in the same
function as regular nodes (delete_instances), but after the check to see
if there are any valid nodes to suspend, this caused an early return
that made that the TPU nodes were never suspended.

Moved the separation of the list between regular nodes and TPU nodes as
well as the call to delete the tpu_instances to the calling function
(suspend_nodes) this way the early return does not affect the invocation
of the delete TPU instances.