Separate packer docker build in two stages
This separation allows to build first a base image and from the
generated base image install a tensorflow version, this makes that all
the tensorflow images share the previous layers, thus reducing the space
used in the GCP registry in a siginicant way. Created a helper script to
generate and push the images.

Disabled also the lustre install in the packer examples vars of docker
build as this was causing issues being compiled inside a container.

Added a helpful note in the ansible role of TPU that states how to get
the list of .whl and .so files in the google artifact repository for
cloud tpu.