Configure slurmd to restart upon-failure
In "config-less" mode for login and compute nodes, slurmd will contact
the controller upon boot and request its configuration. It does not
retry if the controller happens to be down, even momentarily. This
configures SystemD to restart slurmd continually. This is an extremely
simple configuration that is CentOS 7 compatible; more sophisticated
backoff and rate limiting can be applied in later releases of SystemD.