Update configuration for installing lustre
The existing solution breaks whenever new point releases of Rocky linux
are made. The new solution makes the assumption that, after upgrading the
kernel and package for /etc/os-release, that we are running the latest
version of Rocky Linux AND that the latest-release of DDN's packaged
lustre supports the latest version of Rocky linux.