Remove mdformat in favor of pymarkdown
mdformat has a known issue where it will escape brackets (to become
literals) when it encounters GitHub-Flavored Markdown extensions:
https://github.com/executablebooks/mdformat/issues/359#issuecomment-1254998398
It is also fairly agressive in ways that, on net, probably reduce
readability in ASCII form. Swap automation in favor of pymarkdown,
configured identically as in GoogleCloudPlatform/hpc-toolkit.

This hook will be disabled until all markdown files honor the linter