Address failure of compute nodes to join clusters
We have observed real-world failures of VMs to join clusters as compute
nodes due to changing hostnames during boot. Temporarily the VM
identifies its hostname as the FQDN and then reverts (correctly) to the
short hostname. In a significant percentage of boots, the command line
parameters to slurmd are rendered when the hostname is the FQDN.
Removing the explicit parameters allows slurmd restarts to dynamically
detect the hostname, which is typically correct within seconds.