Sync script files from bucket
Compares the checksum of the files contained in the bucket with the ones
in the local directory, and if they do not match, the ones from the
bucket are downloaded locally, these allows to dinamically update the
scripts. To solve a circular dependency the install_custom_scripts has
been moved to util.py.