Allow network_ip input to be null
The current solution requires the network_ip input variable and field in
the additional_networks object to be the empty string when no IP address
is being specified. Otherwise, the `length` function will thrown an
error at plan time. This commit allows these values to be null, which
I believe previous work has demonstrated to be the value that enables
automatic assignment of IP addresses.