Address Terraform provider behavioral change
In https://github.com/hashicorp/terraform-provider-google/commit/bd42980115a2f598c129b9a81fe0a603b0f384d6
Terraform providers 6.2.0 and above make an affirmative comparison of
subnetwork_project in the code to the value in the state. When the
subnetwork provided was a self_link, the subnetwork_project is
automatically modified to be the project from the self_link. With a
default value of the empty string the validation in the commit above
fails for var.subnetwork_project. This validation failure is avoided by
changing the default value to null.